Prepare for multiple multiprocess backends
