+++ prometheus_client/multiprocess/mmap_utils.py
@@ -1,0 +1,9 @@
+import glob
+import os
+
+
+def mmap_cleanup(path, pid):
+    for f in glob.glob(os.path.join(path, 'gauge_livesum_{0}.db'.format(pid))):
+        os.remove(f)
+    for f in glob.glob(os.path.join(path, 'gauge_liveall_{0}.db'.format(pid))):
+        os.remove(f)
